+++ bin/actions/0220022-20.sh
@@ -7,12 +7,10 @@ wmctrl -r :ACTIVE: -b toggle,above
 activeWindowId=`xdotool getactivewindow`
 activeWindowProps=`xprop -id ${activeWindowId} | grep _NET_WM_STATE`
 
-me=$SUDO_USER
-
 if [[ ${activeWindowProps} == *"_NET_WM_STATE_ABOVE"* ]]; then
-    #sudo -u $me echo "Always on top: On" | osd_cat -A right -p top -f -*-*-bold-*-*-*-36-120-*-*-*-*-*-* -cgray -d 1
-    notify-send "Always on top: On" -t 1
+#    notify-send "Always on top: On" -t 1
+    echo "Always on top" | aosd_cat -R white -u 300 --y-offset=40 -B green -b 240 --padding=24 -x 5 -n "Ubuntu 20" -p 1
 else
-    #sudo -u $me echo "Always on top: Off" | osd_cat -A right -p top -f -*-*-bold-*-*-*-36-120-*-*-*-*-*-* -cgray -d 1
-    notify-send "Always on top: Off" -t 1
+#    notify-send "Always on top: Off" -t 1
+    echo "Always on top" | aosd_cat -R white -u 300 --y-offset=40 -B red -b 240 --padding=24 -x 5 -n "Ubuntu 20" -p 1
 fi;