+++ bin/actions/0220022-46.sh
@@ -11,7 +11,9 @@ if [[ "$focusApplication" =~ "chromium-browse" ]] || [[ "$focusApplication" =~ "
   xdotool key Right
   xdotool key shift+ctrl+Left
   xdotool key Ctrl+c
-  notify-send "Copied trailing URL segement" -t 1
+#  notify-send "Copied trailing URL segement" -t 1
+  copied=$(xsel -o)
+  echo "Copied: $copied" | aosd_cat -R white -u 300 --y-offset=40 -B black -b 240 --padding=24 -x 5 -n "Ubuntu 20" -p 1
 else
   if pidof -s chromium-browse >/dev/null; then
     wmctrl -a Chromium  # Activate Chromium window