+++ bin/actions/1010101-108.sh
@@ -4,7 +4,7 @@
 
 amixer -D pulse sset Master 5%- &> /dev/null
 
-me=$SUDO_USER
 currentVolume=`amixer -D pulse sget Master | grep 'Front Left: Playback' | grep -oE "[0-9]{1,3}%"`
-#sudo -u $me echo "Volume lower ($currentVolume)" | osd_cat -A right -p top -f -*-*-bold-*-*-*-36-120-*-*-*-*-*-* -cgray -d 1
-notify-send "Volume lower ($currentVolume)" -t 1
+#notify-send "Volume lower ($currentVolume)" -t 1
+echo "Volume--" | aosd_cat -R white -u 300 --y-offset=40 -B black -b 240 --padding=24 -x 5 -n "Ubuntu 20" -p 1
+xdotool key Up