+++ CHANGELOG.md
@@ -5,6 +5,7 @@ V0.2.2 - Not released yet
 -------------------------
 * Add browser action: ◢ + Backspace = Remove trailing segment from current URL
 * Add browser action: ◢ + C = Copy trailing segment of URL
+* Change: Use aosd-cat instead of notify-send for notifications
 
 V0.2.1 - 2020/12/13
 -------------------